Change to f32 for better performance of SIMD.

+++ src/geoalg/f32_math/convolution_kernel.rs
@@ -2,6 +2,6 @@ struct ConvolutionKernel {
     rows: usize,
     columns: usize,
 
-    data: Vec<f64>
+    data: Vec<f32>
 }
